Move `get_dtype_identifier` method into the `ArrayBackend` class.
This is another method that was in `TensorSpace` out of necessity, but really does
not have anything to do with ODL / mathematics but only with backend-specific
ways of handling dtypes.
